feat: start VMs on startup
