+++ nilcc-agent/src/services/mod.rs
@@ -1,3 +1,4 @@
 pub mod disk;
 pub mod proxy;
+pub mod vm;
 pub mod workload;
