+++ nilcc-agent/src/services/workload.rs
@@ -1,20 +1,16 @@
 use crate::{
-    clients::qemu::{HardDiskFormat, HardDiskSpec, VmSpec},
-    config::{CvmConfig, CvmFiles},
-    iso::{ApplicationMetadata, ContainerMetadata, EnvironmentVariable, IsoSpec},
     repositories::workload::{Workload, WorkloadRepository, WorkloadRepositoryError},
     resources::{GpuAddress, SystemResources},
     routes::workloads::create::CreateWorkloadRequest,
     services::{
-        disk::{DiskService, DockerComposeHash},
-        proxy::ProxyService,
+        proxy::{ProxiedVm, ProxyService},
+        vm::{StartVmError, VmService},
     },
-    workers::scheduler::VmSchedulerHandle,
 };
 use async_trait::async_trait;
-use std::{collections::BTreeSet, fmt, io, ops::Range, path::PathBuf};
+use std::{collections::BTreeSet, io, ops::Range};
 use strum::EnumDiscriminants;
-use tokio::{fs, sync::Mutex};
+use tokio::sync::Mutex;
 use tracing::info;
 use uuid::Uuid;
 
@@ -35,6 +31,12 @@ pub enum CreateWorkloadError {
 
     #[error("workload already exists")]
     AlreadyExists,
+}
+
+impl From<StartVmError> for CreateWorkloadError {
+    fn from(e: StartVmError) -> Self {
+        Self::Internal(e.to_string())
+    }
 }
 
 impl From<WorkloadRepositoryError> for CreateWorkloadError {
@@ -67,14 +69,11 @@ impl From<WorkloadRepositoryError> for DeleteWorkloadError {
 }
 
 pub struct WorkloadServiceArgs {
-    pub state_path: PathBuf,
-    pub disk_service: Box<dyn DiskService>,
-    pub scheduler: Box<dyn VmSchedulerHandle>,
+    pub vm_service: Box<dyn VmService>,
     pub repository: Box<dyn WorkloadRepository>,
     pub proxy_service: Box<dyn ProxyService>,
     pub resources: SystemResources,
     pub open_ports: Range<u16>,
-    pub cvm_config: CvmConfig,
 }
 
 struct AvailableResources {
@@ -107,31 +106,21 @@ pub enum CreateServiceError {
 
     #[error("database: {0}")]
     Database(#[from] WorkloadRepositoryError),
+
+    #[error("staring existing workload: {0}")]
+    StartWorkload(#[from] StartVmError),
 }
 
 pub struct DefaultWorkloadService {
-    state_path: PathBuf,
     repository: Box<dyn WorkloadRepository>,
-    scheduler: Box<dyn VmSchedulerHandle>,
-    disk_service: Box<dyn DiskService>,
+    vm_service: Box<dyn VmService>,
     proxy_service: Box<dyn ProxyService>,
     resources: Mutex<AvailableResources>,
-    cvm_config: CvmConfig,
 }
 
 impl DefaultWorkloadService {
     pub async fn new(args: WorkloadServiceArgs) -> Result<Self, CreateServiceError> {
-        let WorkloadServiceArgs {
-            state_path,
-            disk_service,
-            scheduler,
-            repository,
-            proxy_service,
-            resources,
-            open_ports,
-            cvm_config,
-        } = args;
-        fs::create_dir_all(&state_path).await.map_err(CreateServiceError::StateDirectory)?;
+        let WorkloadServiceArgs { vm_service, repository, proxy_service, resources, open_ports } = args;
 
         let workloads = repository.list().await?;
         let mut gpus: BTreeSet<_> = resources.gpus.iter().flat_map(|g| g.addresses.iter().cloned()).collect();
@@ -141,9 +130,9 @@ impl DefaultWorkloadService {
         let mut disk_space_gb = resources.available_disk_space_gb();
         for workload in workloads {
             let workload_id = workload.id;
-            for gpu in workload.gpus {
-                if !gpus.remove(&gpu) {
-                    return Err(CreateServiceError::CommittedGpuMissing(workload_id, gpu));
+            for gpu in &workload.gpus {
+                if !gpus.remove(gpu) {
+                    return Err(CreateServiceError::CommittedGpuMissing(workload_id, gpu.clone()));
                 }
             }
             for port in [workload.proxy_http_port, workload.proxy_https_port] {
@@ -155,11 +144,13 @@ impl DefaultWorkloadService {
             memory_mb = memory_mb.checked_sub(workload.memory_mb).ok_or(CreateServiceError::OvercommittedMemory)?;
             disk_space_gb =
                 disk_space_gb.checked_sub(workload.disk_space_gb).ok_or(CreateServiceError::OvercommittedDiskSpace)?;
+            info!("Starting existing workload {workload_id}");
+            vm_service.start_vm(workload).await?;
         }
         let gpus = gpus.into_iter().collect();
         let ports = ports.into_iter().collect();
         let resources = AvailableResources { cpus, gpus, ports, memory_mb, disk_space_gb }.into();
-        Ok(Self { state_path, scheduler, disk_service, repository, proxy_service, resources, cvm_config })
+        Ok(Self { vm_service, repository, proxy_service, resources })
     }
 
     fn build_workload(&self, request: CreateWorkloadRequest, resources: &AvailableResources) -> Workload {
@@ -191,74 +182,6 @@ impl DefaultWorkloadService {
             proxy_https_port: ports[1],
         }
     }
-
-    fn create_vm_spec(
-        &self,
-        workload: &Workload,
-        iso_path: PathBuf,
-        state_disk_path: PathBuf,
-        docker_compose_hash: DockerComposeHash,
-    ) -> VmSpec {
-        let CvmFiles { base_disk, kernel, verity_root_hash, verity_disk } =
-            if workload.gpus.is_empty() { &self.cvm_config.cpu } else { &self.cvm_config.gpu };
-        let kernel_args =
-            KernelArgs { filesystem_root_hash: verity_root_hash, docker_compose_hash: &docker_compose_hash.0 };
-        VmSpec {
-            cpu: workload.cpus,
-            ram_mib: workload.memory_mb,
-            hard_disks: vec![
-                HardDiskSpec { path: base_disk.clone(), format: HardDiskFormat::Qcow2 },
-                HardDiskSpec { path: verity_disk.clone(), format: HardDiskFormat::Raw },
-                HardDiskSpec { path: state_disk_path, format: HardDiskFormat::Raw },
-            ],
-            cdrom_iso_path: Some(iso_path),
-            gpus: workload.gpus.clone(),
-            port_forwarding: vec![(workload.proxy_http_port, 80), (workload.proxy_https_port, 443)],
-            bios_path: Some(self.cvm_config.bios.clone()),
-            initrd_path: Some(self.cvm_config.initrd.clone()),
-            kernel_path: Some(kernel.clone()),
-            kernel_args: Some(kernel_args.to_string()),
-            display_gtk: false,
-            enable_cvm: true,
-        }
-    }
-
-    async fn create_state_disk(&self, workload: &Workload) -> Result<PathBuf, CreateWorkloadError> {
-        let disk_name = format!("{}.raw", workload.id);
-        let disk_path = self.state_path.join(disk_name);
-        self.disk_service
-            .create_disk(&disk_path, HardDiskFormat::Raw, workload.disk_space_gb)
-            .await
-            .map_err(|e| CreateWorkloadError::Internal(format!("failed to create state disk: {e}")))?;
-        Ok(disk_path)
-    }
-
-    async fn create_application_iso(
-        &self,
-        workload: &Workload,
-    ) -> Result<(PathBuf, DockerComposeHash), CreateWorkloadError> {
-        let iso_name = format!("{}.iso", workload.id);
-        let iso_path = self.state_path.join(iso_name);
-        let environment_variables =
-            workload.env_vars.iter().map(|(name, value)| EnvironmentVariable::new(name, value)).collect();
-        let spec = IsoSpec {
-            docker_compose_yaml: workload.docker_compose.clone(),
-            metadata: ApplicationMetadata {
-                hostname: "UNKNOWN".into(),
-                api: ContainerMetadata {
-                    container: workload.public_container_name.clone(),
-                    port: workload.public_container_port,
-                },
-            },
-            environment_variables,
-        };
-        let docker_compose_hash = self
-            .disk_service
-            .create_application_iso(&iso_path, spec)
-            .await
-            .map_err(|e| CreateWorkloadError::Internal(format!("failed to create ISO: {e}")))?;
-        Ok((iso_path, docker_compose_hash))
-    }
 }
 
 #[async_trait]
@@ -287,18 +210,12 @@ impl WorkloadService for DefaultWorkloadService {
         let workload = self.build_workload(request, &resources);
         let id = workload.id;
         info!("Storing workload {id} in database");
-        // TODO: this should be a create, change once everything is migrated to new api
         self.repository.create(workload.clone()).await?;
 
-        info!("Creating disks for VM {id}");
-        let (iso_path, docker_compose_hash) = self.create_application_iso(&workload).await?;
-        let state_disk = self.create_state_disk(&workload).await?;
-        let spec = self.create_vm_spec(&workload, iso_path, state_disk, docker_compose_hash);
-
         info!("Scheduling VM {id}");
-        let socket_path = self.state_path.join(format!("{}.sock", workload.id));
-        self.scheduler.start_vm(workload.id, spec, socket_path).await;
-        self.proxy_service.start_vm_proxy((&workload).into()).await;
+        let proxied_vm = ProxiedVm::from(&workload);
+        self.vm_service.start_vm(workload).await?;
+        self.proxy_service.start_vm_proxy(proxied_vm).await;
 
         resources.cpus -= cpus;
         resources.gpus.drain(0..gpus);
@@ -315,20 +232,8 @@ impl WorkloadService for DefaultWorkloadService {
         info!("Deleting workload: {id}");
         self.repository.delete(id).await?;
         self.proxy_service.stop_vm_proxy(id).await;
-        self.scheduler.stop_vm(id).await;
+        self.vm_service.stop_vm(id).await;
         Ok(())
-    }
-}
-
-struct KernelArgs<'a> {
-    docker_compose_hash: &'a str,
-    filesystem_root_hash: &'a str,
-}
-
-impl fmt::Display for KernelArgs<'_> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let Self { docker_compose_hash, filesystem_root_hash } = self;
-        write!(f, "panic=-1 root=/dev/sda2 verity_disk=/dev/sdb verity_roothash={filesystem_root_hash} state_disk=/dev/sdc docker_compose_disk=/dev/sr0 docker_compose_hash={docker_compose_hash}")
     }
 }
 
@@ -339,25 +244,21 @@ mod tests {
         repositories::workload::MockWorkloadRepository,
         resources::Gpus,
         services::{
-            disk::MockDiskService,
             proxy::{MockProxyService, ProxiedVm},
+            vm::MockVmService,
         },
-        workers::scheduler::MockVmSchedulerHandle,
     };
     use mockall::predicate::eq;
     use rstest::rstest;
     use uuid::Uuid;
 
     struct Builder {
-        scheduler: MockVmSchedulerHandle,
+        vm_service: MockVmService,
         repository: MockWorkloadRepository,
-        disk_service: MockDiskService,
         proxy_service: MockProxyService,
         resources: SystemResources,
         open_ports: Range<u16>,
         existing_workloads: Vec<Workload>,
-        cvm_config: CvmConfig,
-        state_path: PathBuf,
     }
 
     impl Builder {
@@ -370,27 +271,14 @@ mod tests {
         }
 
         async fn try_build(self) -> Result<DefaultWorkloadService, CreateServiceError> {
-            let Self {
-                scheduler,
-                mut repository,
-                disk_service,
-                proxy_service,
-                resources,
-                open_ports,
-                existing_workloads,
-                cvm_config,
-                state_path,
-            } = self;
+            let Self { vm_service, mut repository, proxy_service, resources, open_ports, existing_workloads } = self;
             repository.expect_list().return_once(move || Ok(existing_workloads));
             let args = WorkloadServiceArgs {
-                scheduler: Box::new(scheduler),
+                vm_service: Box::new(vm_service),
                 repository: Box::new(repository),
-                disk_service: Box::new(disk_service),
                 proxy_service: Box::new(proxy_service),
                 resources,
                 open_ports,
-                cvm_config,
-                state_path,
             };
             DefaultWorkloadService::new(args).await
         }
@@ -399,9 +287,8 @@ mod tests {
     impl Default for Builder {
         fn default() -> Self {
             Self {
-                scheduler: Default::default(),
+                vm_service: Default::default(),
                 repository: Default::default(),
-                disk_service: Default::default(),
                 proxy_service: Default::default(),
                 resources: SystemResources {
                     hostname: "foo".into(),
@@ -415,23 +302,6 @@ mod tests {
                 },
                 open_ports: 100..200,
                 existing_workloads: Default::default(),
-                cvm_config: CvmConfig {
-                    initrd: "/initrd".into(),
-                    bios: "/bios".into(),
-                    cpu: CvmFiles {
-                        base_disk: "/cpu/base-disk".into(),
-                        kernel: "/cpu/kernel".into(),
-                        verity_disk: "/cpu/verity-disk".into(),
-                        verity_root_hash: "cpu-root-hash".into(),
-                    },
-                    gpu: CvmFiles {
-                        base_disk: "/gpu/base-disk".into(),
-                        kernel: "/gpu/kernel".into(),
-                        verity_disk: "/gpu/verity-disk".into(),
-                        verity_root_hash: "gpu-root-hash".into(),
-                    },
-                },
-                state_path: PathBuf::from("/tmp"),
             }
         }
     }
@@ -513,7 +383,9 @@ mod tests {
             proxy_https_port: 1001,
             ..make_workload()
         };
-        builder.existing_workloads = vec![workload];
+        builder.existing_workloads = vec![workload.clone()];
+        // We should start this VM as part of the startup process
+        builder.vm_service.expect_start_vm().with(eq(workload)).once().return_once(move |_| Ok(()));
 
         let service = builder.build().await;
         let resources = service.resources.lock().await;
@@ -557,35 +429,7 @@ mod tests {
             proxy_https_port: 101,
         };
         let mut builder = Builder::default();
-        let docker_compose_hash = "deadbeef";
         let id = workload.id;
-        let base_path = builder.state_path.display();
-        let iso_path = PathBuf::from(format!("{base_path}/{id}.iso"));
-        let state_disk_path = PathBuf::from(format!("{base_path}/{id}.raw"));
-        let socket_path = PathBuf::from(format!("{base_path}/{id}.sock"));
-        let kernel_args = KernelArgs {
-            docker_compose_hash: docker_compose_hash,
-            filesystem_root_hash: &builder.cvm_config.gpu.verity_root_hash,
-        }
-        .to_string();
-        let spec = VmSpec {
-            cpu: 1,
-            ram_mib: 1024,
-            hard_disks: vec![
-                HardDiskSpec { path: "/gpu/base-disk".into(), format: HardDiskFormat::Qcow2 },
-                HardDiskSpec { path: "/gpu/verity-disk".into(), format: HardDiskFormat::Raw },
-                HardDiskSpec { path: state_disk_path.clone(), format: HardDiskFormat::Raw },
-            ],
-            cdrom_iso_path: Some(iso_path.clone()),
-            gpus: vec!["addr1".into()],
-            port_forwarding: vec![(workload.proxy_http_port, 80), (workload.proxy_https_port, 443)],
-            bios_path: Some("/bios".into()),
-            initrd_path: Some("/initrd".into()),
-            kernel_path: Some("/gpu/kernel".into()),
-            kernel_args: Some(kernel_args),
-            display_gtk: false,
-            enable_cvm: true,
-        };
 
         let expected_cpus = builder.resources.available_cpus() - request.cpus as u32;
         let expected_memory = builder.resources.available_memory_mb() - request.memory_mb;
@@ -594,25 +438,11 @@ mod tests {
         builder.open_ports = 100..200;
         builder.resources.gpus = Some(Gpus::new("H100", ["addr1".into()]));
         builder.repository.expect_create().with(eq(workload.clone())).once().return_once(|_| Ok(()));
-        builder
-            .disk_service
-            .expect_create_application_iso()
-            .return_once(move |_, _| Ok(DockerComposeHash(docker_compose_hash.into())));
-        builder
-            .disk_service
-            .expect_create_disk()
-            .with(eq(state_disk_path), eq(HardDiskFormat::Raw), eq(1))
-            .return_once(move |_, _, _| Ok(()));
-        builder
-            .scheduler
-            .expect_start_vm()
-            .with(eq(workload.id), eq(spec), eq(socket_path))
-            .once()
-            .return_once(|_, _, _| ());
+        builder.vm_service.expect_start_vm().with(eq(workload)).once().return_once(|_| Ok(()));
         builder
             .proxy_service
             .expect_start_vm_proxy()
-            .with(eq(ProxiedVm { id: workload.id, http_port: 100, https_port: 101 }))
+            .with(eq(ProxiedVm { id, http_port: 100, https_port: 101 }))
             .return_once(move |_| ());
 
         let service = builder.build().await;
