+++ nilcc-agent/src/workers/mod.rs
@@ -1,2 +1,1 @@
-pub mod scheduler;
 pub(crate) mod vm;
